+++ components/ContentstackRichText/BasicUsageBlock/IconDo.tsx
@@ -1,8 +1,8 @@
 function IconDo(props) {
   return (
     <svg
-      width={36}
-      height={36}
+      width={24}
+      height={24}
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
       viewBox="0 0 36 36"
