+++ components/ContentstackRichText/BasicUsageBlock/index.tsx
@@ -1,0 +1,3 @@
+import BasicUsageBlock from './BasicUsageBlock';
+
+export default BasicUsageBlock;
